Made sure to pass longs in to curl_easy_setopt where necessary in the libtest code.

+++ tests/libtest/lib552.c
@@ -5,7 +5,7 @@
  *                            | (__| |_| |  _ <| |___
  *                             \___|\___/|_| \_\_____|
  *
- * $Id: lib552.c,v 1.3 2008-02-27 09:06:15 bagder Exp $
+ * $Id: lib552.c,v 1.4 2008-05-22 21:49:53 danf Exp $
  *
  * argv1 = URL
  * argv2 = proxy with embedded user+password
@@ -158,17 +158,17 @@ int test(char *URL)
     curl_easy_setopt(curl, CURLOPT_DEBUGFUNCTION, my_trace);
     curl_easy_setopt(curl, CURLOPT_DEBUGDATA, &config);
     /* the DEBUGFUNCTION has no effect until we enable VERBOSE */
-    curl_easy_setopt(curl, CURLOPT_VERBOSE, 1);
+    curl_easy_setopt(curl, CURLOPT_VERBOSE, 1L);
 
     /* setup repeated data string */
     for (i=0; i < sizeof data; ++i)
         data[i] = fill[i % sizeof fill];
 
     /* Post */
-    curl_easy_setopt(curl, CURLOPT_POST, 1);
+    curl_easy_setopt(curl, CURLOPT_POST, 1L);
 
     /* Setup read callback */
-    curl_easy_setopt(curl, CURLOPT_POSTFIELDSIZE, sizeof data);
+    curl_easy_setopt(curl, CURLOPT_POSTFIELDSIZE, (long) sizeof data);
     curl_easy_setopt(curl, CURLOPT_READFUNCTION, read_callback);
 
     /* Write callback */
@@ -182,7 +182,7 @@ int test(char *URL)
     curl_easy_setopt(curl, CURLOPT_URL, URL);
 
     /* Accept any auth. But for this bug configure proxy with DIGEST, basic might work too, not NTLM */
-    curl_easy_setopt(curl, CURLOPT_PROXYAUTH, CURLAUTH_ANY);
+    curl_easy_setopt(curl, CURLOPT_PROXYAUTH, (long)CURLAUTH_ANY);
 
     res = curl_easy_perform(curl);
     fprintf(stderr, "curl_easy_perform = %d\n", (int)res);
